rename reducer/compile as reducer/compiler
This commit renames reducer/compile as "compiler" to have consistent
naming convention for compiler abstractons in the zq code base.
